move something to devices
