+++ packages/react-client/src/devices/DeviceManager.ts
@@ -6,17 +6,17 @@ import type {
   Media,
   MediaManager,
   MediaStatus,
-} from "./types/internal";
+} from "../types/internal";
 
-import { prepareMediaTrackConstraints } from "./constraints";
+import { prepareMediaTrackConstraints } from "../constraints";
 
 import EventEmitter from "events";
 import type TypedEmitter from "typed-emitter";
-import type { PersistLastDeviceHandlers, TrackMiddleware } from "./types/public";
-import { MiddlewareManager } from "./devices/MiddlewareManager";
-import { createStorageConfig } from "./utils/localStorage";
-import { setupOnEndedCallback } from "./utils/track";
-import { parseUserMediaError } from "./utils/errors";
+import type { PersistLastDeviceHandlers, TrackMiddleware } from "../types/public";
+import { MiddlewareManager } from "./MiddlewareManager";
+import { createStorageConfig } from "../utils/localStorage";
+import { setupOnEndedCallback } from "../utils/track";
+import { parseUserMediaError } from "../utils/errors";
 
 export type DeviceType = "audio" | "video";
