+++ packages/react-client/src/devices/mediaInitializer.ts
@@ -1,6 +1,6 @@
-import { prepareConstraints } from "./constraints";
-import type { CurrentDevices, DeviceError } from "./types/internal";
-import { NOT_FOUND_ERROR, OVERCONSTRAINED_ERROR, PERMISSION_DENIED, UNHANDLED_ERROR } from "./utils/errors";
+import { prepareConstraints } from "../constraints";
+import type { CurrentDevices, DeviceError } from "../types/internal";
+import { NOT_FOUND_ERROR, OVERCONSTRAINED_ERROR, PERMISSION_DENIED, UNHANDLED_ERROR } from "../utils/errors";
 
 type AudioVideo<T> = { audio: T; video: T };
