+++ packages/react-client/src/hooks/internal/device/useDeviceApi.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import type { TrackManager } from "../../../types/internal";
 import type { Device } from "../../../types/public";
-import type { DeviceManager } from "../../../DeviceManager";
+import type { DeviceManager } from "../../../devices/DeviceManager";
 import { useDeviceManager } from "./useDeviceManager";
 
 type DeviceApiDependencies = {
